backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
adjusttext==0.7.3
aiohttp==3.8.1; python_version >= '3.6'
aiosignal==1.2.0; python_version >= '3.6'
asgiref==3.5.0; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
automat==20.2.0
base58==2.1.1; python_version >= '3.5'
bitarray==1.2.2
cachecontrol==0.12.10; python_version >= '3.6'
cachetools==5.0.0; python_version ~= '3.7'
certifi==2021.10.8
charset-normalizer==2.0.11; python_version >= '3.5'
click==8.0.3; python_version >= '3.6'
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0; python_version >= '3.6'
cytoolz==0.11.2; python_version >= '3.5'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
django-cors-headers==3.7.0
django-extensions==3.1.5
django-filter==21.1
django-redis==5.2.0
django-storages[google]==1.12.3
django-waffle==2.3.0
django==3.2.11
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
drf-yasg==1.20.0
eth-abi==2.1.1; python_version >= '3.6' and python_version < '4.0'
eth-account==0.5.7; python_version >= '3.6' and python_version < '4.0'
eth-hash[pycryptodome]==0.3.2; python_version >= '3.5' and python_version < '4.0'
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1; python_version >= '3.6' and python_version < '4.0'
eth-typing==2.3.0; python_version >= '3.5' and python_version < '4.0'
eth-utils==1.10.0; python_version >= '3.5' and python_version < '4.0' and python_full_version != '3.5.2'
firebase-admin==5.2.0
fonttools==4.29.0; python_version >= '3.7'
frozenlist==1.3.0; python_version >= '3.7'
google-api-core[grpc]==2.4.0; platform_python_implementation != 'PyPy'
google-api-python-client==2.36.0; python_version >= '3.6'
google-auth-httplib2==0.1.0
google-auth==2.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-core==2.2.2; python_version >= '3.6'
google-cloud-firestore==2.3.4; platform_python_implementation != 'PyPy'
google-cloud-storage==2.1.0; python_version >= '3.7'
google-crc32c==1.3.0; python_version >= '3.6'
google-resumable-media==2.1.0; python_version >= '3.6'
googleapis-common-protos==1.54.0; python_version >= '3.6'
grpcio-status==1.43.0
grpcio==1.43.0
h11==0.13.0; python_version >= '3.6'
hexbytes==0.2.2; python_version >= '3.6' and python_version < '4.0'
httplib2==0.20.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
hyperlink==21.0.0
idna==3.3; python_version >= '3'
incremental==21.3.0
inflection==0.5.1; python_version >= '3.5'
ipfshttpclient==0.8.0a2; python_full_version >= '3.6.2' and python_full_version not in '3.7.0, 3.7.1'
itypes==1.2.0
jinja2==3.0.3; python_version >= '3.6'
joblib==1.1.0; python_version >= '3.6'
jsonschema==3.2.0
kiwisolver==1.3.2; python_version >= '3.7'
lru-dict==1.1.7
many==0.6.9
markupsafe==2.0.1; python_version >= '3.6'
matplotlib==3.5.1; python_version >= '3.7'
msgpack==1.0.3
multiaddr==0.0.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
multidict==6.0.2; python_version >= '3.7'
netaddr==0.8.0
numpy==1.22.1; python_version >= '3.8'
packaging==21.3; python_version >= '3.6'
pandas==1.4.0; python_version >= '3.8'
parsimonious==0.8.1
patsy==0.5.2
pillow==9.0.0
proto-plus==1.19.9; python_version >= '3.6'
protobuf==3.19.4; python_version >= '3.5'
psycopg2-binary==2.9.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycryptodome==3.14.0
pyjwt==2.3.0; python_version >= '3.6'
pyparsing==3.0.7; python_version >= '3.1'
pyrsistent==0.18.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
qrcode==7.3.1
redis==4.1.2; python_version >= '3.6'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rlp==2.0.1
rsa==4.8; python_version >= '3.6'
ruamel.yaml.clib==0.2.6; python_version < '3.11' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.20; python_version >= '3'
scikit-learn==1.0.2; python_version >= '3.7'
scipy==1.7.3; python_version < '3.11' and python_version >= '3.7'
seaborn==0.11.2; python_version >= '3.6'
setuptools==60.6.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sklearn==0.0
sqlparse==0.4.2; python_version >= '3.5'
statsmodels==0.13.1; python_version >= '3.7'
threadpoolctl==3.0.0; python_version >= '3.6'
toolz==0.11.2; python_version >= '3.5'
tqdm==4.62.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
twisted==21.7.0; python_full_version >= '3.6.7'
txamqp==0.8.2
typing-extensions==4.0.1; python_version >= '3.6'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.8
username==2.1.0
uvicorn==0.17.1
varint==1.0.2
web3==5.26.0
websockets==9.1; python_full_version >= '3.6.1'
wrapt==1.13.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
yarl==1.7.2; python_version >= '3.6'
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
